+++ qiskit/__init__.py
@@ -31,7 +31,7 @@ from ._measure import Measure
 import qiskit.extensions.standard
 import qiskit.extensions.quantum_initializer
 
-from ._result import Result
+from .result import Result
 
 from .wrapper._wrapper import (
     available_backends, local_backends, remote_backends,