+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ Changed
     - introduce the ``qiskit.qobj`` module. (#589, #655)
     - update the ``Qobj`` JSON schema. (#668, #677, #703, #709)
     - update the local simulators for accepting ``Qobj`` as input. (#667)
+    - update the ``Result`` class. (#773)
 - Use ``get_status_job()`` for checking IBMQJob status. (#641)
 - Q network hub/group/project credentials replaced by new url format. (#740)
 - ``Jobs`` API simplification and breaking changes. (#686)